Move code to create ApolloReporter
